feat: Complete the menu maker function
In this commit, the menu maker function, menuItemsMaker was refactored to make
use of a forEach loop instead of a for loop. when an array of menu items
is passed into the menu maker function, it loops through the array and creates
list items for each item of the array.

the created li item is then appended to the un-ordered list, ul.

An event listener was added to the hamburge menu image of class .menu-button.
when this menu image is clicked, it toggles a css class menu--open on the
menu_button variable in the menu.js file